Refactor for clarity
